+++ Source/EquationElements/NumericValue.hpp
@@ -1,6 +1,6 @@
 /**
  * NumericValue.hpp
- * Purpose: A class to represent an integer value in an equation.
+ * Purpose: A class to represent a float value in an equation.
  * Author: C. Cooper
  * Date Created: 2023-11-28
 */
@@ -12,16 +12,16 @@
 #include <string>
 
 /**
- * Represents a constant integer value in an equation.
+ * Represents a constant floating point value in an equation.
 */
 class NumericValue : public EquationElement
 {
     private:
-    int value;
+    float value;
 
     public:
-    NumericValue(int);
-    int getValue(EquationElement*, EquationElement*) override;
+    NumericValue(float);
+    float getValue(EquationElement*, EquationElement*) override;
     int getPriority() override;
     std::string getSymbol() override;
 };
